backend/app/requirements.txt
pymongo 
flask 
flask-security
flask-admin
flask-mongoengine
flask_restful
psutil
requests
ipython==5.3.0
PyJWT